examples/requirements.txt
pytorch-lightning 
git+git://github.com/kschlegel/DatasetLoader.git@719908d828155dd12104af44a4b14fa5c3b910f4
